[structure] move useXML to a inner component
